requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-mcp-adapters>=0.1.0
fastmcp>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=13.0
